Update hyperv driver so that:
When the driver is created a handle to the FileMapping for the memory for this sandbox.

Uses that handle as an argument to the get_surrogate_process function on the SurrogateProcessManager

Removes the code that copies the Sandbox memory to and from the surrogate process memory each time the VM is run.

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>