Add clarification on why we must take GuestSharedMemory::lock
Previously, the code comments were not sufficiently clear as to why
apparently-useless locks during the execution of the guest are
necessary for the invariants of the GuestSharedMemory/HostSharedMemory
split.  This led to confusion as to whether it would be a good idea to
remove them, so this commit clarifies that there is a good reason for
them to be there.

Signed-off-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>