Allow explicit setting of max guest log level, check RUST_LOG to set guest log_level if not set
The guest log level impacts the amount of logs produced by the guest, its presently set by examining the value of max_log_level in the host.

This may not be appropriate since logging in the guest is expensive and it may cause logs to be emitted from the guest but ignored in the host. This change introduces a function on an uninitializedsandbox to set the level explicitly, if this is not set the env var RUST_LOG is parsed to determine the log level so that logs  are only produced if they are going to be consumed.  If RUST_LOG is not set then the max guest log level is set to error.

Signed-off-by: Simon Davies <simongdavies@users.noreply.github.com>