[common/{mem,outb},guest/*,host/{mem,func,sandbox,tests}] removed guest panic context mem region
We had a guest panic context region to store the panic msg. Now, instead of storing info in shared mem, we
use debug_print. To still have some exception info  when using OutBAction::Abort, we pass parse the
exception name, which is more readable than before and provides information that is useful as illustrated
in some of the integration tests.

Signed-off-by: danbugs <danilochiarlone@gmail.com>