[guest/exceptions] changed the ABI of exception related logic
We used the sysv64 ABI for exception stuff. This ABI is not available for every target so having it there
unconditionally makes the hyperlight_guest lib not be compilable to said targets. This commit changes the ABI to "C"
(i.e., smt more commonly used) and gates it behind a x86_64 cfg as the assembly code is dependent on 64-bit
architectures.

Signed-off-by: danbugs <danilochiarlone@gmail.com>