[host] added GuestEnvironment struct
The GuestEnvironment struct contains two blobs of data. One identifiable as a guest binary,
and one undifferentiated guest blob. This GuestEnvironment is now used to create a new
sandbox in place of just a guest binary. There are TryFrom impls to be able to convert from
a guest binary to a GuestEnvironment, so this isn't a breaking change.

Signed-off-by: danbugs <danilochiarlone@gmail.com>