Updates for wasmparser apis
This was a bit more involved due to a big change in the way function return types are handled.  The orginal change was done in https://github.com/WebAssembly/component-model/pull/368 and put initially put behind a feature flag in https://github.com/bytecodealliance/wasm-tools/pull/1670. In 0.226.0 version of wasmparser this feature flag was removed simplify the return results.  This in turn simplifies the handling of the results here as well.

Signed-off-by: James Sturtevant <jsturtevant@gmail.com>