[hyperlight_guest_bin] Disable red zones when compiling C
Because we don't presently switch stacks on interrupt entry, we cannot
support a red zone in the ABI. rustc's x86_64-unknown-none target
already does not use a redzone, but clang's x86_64-unknown-none-linux
target, which we use for compiling C code, does use a red zone by
default.  This commit modifies the clang options that we use to remove
uses of a red zone from generated code.

Signed-off-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>