[hyperlight_guest_bin] Save/restore x87 & SSE regs on context switch
We unfortunately don't have a good way of making sure that code run in
exception contexts doesn't use floating point and SSE
instructions. This commit simply always saves and restores the
relevant registers as a workaround. We should at some point look into
the alternative of ensuring that simd instructions aren't used in code
reachable from an exception handler, and see if that actually improves
performance.

Signed-off-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>