[hyperlight_host] mshv: dump core on unknown HV message
Usually, an unknown HV message is the result of a double fault or
similar: something going wrong in the guest. This commit ensures that a
core file is generated when this happens and the crashdump feature is
enabled.

Signed-off-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>

[hyperlight_host] Expose MemoryRegion structure

It is a convenient tuple of information to use when mapping a new region
into the sandbox.

Signed-off-by: Lucy Menon <168595099+syntactically@users.noreply.github.com>