fix(guest-bin): align user memory allocations (#753)
* fix(guest-bin): align user memory allocations

The guest allocation wrapper aligned only the total memory block but did
not guarantee that the pointer returned to the user was itself properly
aligned. This patch attempts to fix that.

It also adds alloc_aligned function which would be tricky to implement
outside the guest but will make implementing other posix like allocation
API much easier.

Signed-off-by: Tomasz Andrzejak <andreiltd@gmail.com>

* Add comment on alloc_helper invariants

Signed-off-by: Tomasz Andrzejak <andreiltd@gmail.com>

---------

Signed-off-by: Tomasz Andrzejak <andreiltd@gmail.com>